DDF-4509 federation -> metacard-federation.js
