--- ui/packages/catalog-ui-search/src/main/webapp/react-component/container/workspace-interactions/federation.js
@@ -1,13 +1,0 @@
-import { Model } from 'backbone'
-
-module.exports = new (Model.extend({
-  initialize: function() {
-    this.set('federation', 'enterprise')
-  },
-  setFederation: function(federation) {
-    this.set('federation', federation)
-  },
-  getFederation: function() {
-    return this.get('federation')
-  },
-}))()
